requirements.txt
pyinstaller
flask
cased-kit
